src/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
Flask == 1.1.2
Pillow == 8.1.0
pandas == 1.2.1
python_docx == 0.8.10
requests == 2.25.1
